projects/intro-to-vector-dbs/pyproject.toml
[project]
name = "intro-to-vector-dbs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langchain-pinecone>=0.2.12",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
]
